rag/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-multipart==0.0.6
chromadb==0.4.22
langchain==0.1.9
langchain-google-genai>=0.0.11
google-generativeai>=0.3.2
pypdf==4.0.1
python-dotenv==1.0.1
sentence-transformers==2.5.1 